Refactor: 한 파일에 모여져있던 atoms를 각각의 파일로 분할 (#12)
