+++ src/atoms/docCountByDateState.ts
@@ -1,0 +1,6 @@
+import { atom } from 'recoil';
+
+export const docCountByDateState = atom<{ [key: string]: number }>({
+  key: 'documentCount',
+  default: {},
+});
