+++ src/atoms/docIdsByDateState.ts
@@ -1,0 +1,6 @@
+import { atom } from 'recoil';
+
+export const docIdsByDateState = atom<{ [key: string]: string[] }>({
+  key: 'docIdsByDate',
+  default: {},
+});
