+++ src/atoms/docIdsState.ts
@@ -1,0 +1,6 @@
+import { atom } from 'recoil';
+
+export const docIdsState = atom<string[]>({
+  key: 'docIds',
+  default: [],
+});
