+++ src/atoms/isOffCanvasOpenState.ts
@@ -1,0 +1,6 @@
+import { atom } from 'recoil';
+
+export const isOffCanvasOpenState = atom({
+  key: 'toggleMenu',
+  default: true,
+});
