+++ src/atoms/isWideState.ts
@@ -1,0 +1,6 @@
+import { atom } from 'recoil';
+
+export const isWideState = atom({
+  key: 'isWide',
+  default: window.innerWidth > 1024,
+});
