+++ src/atoms/lodaingState.ts
@@ -1,0 +1,6 @@
+import { atom } from 'recoil';
+
+export const loadingState = atom({
+  key: 'loading',
+  default: false,
+});
