+++ src/atoms/myListModalState.ts
@@ -1,0 +1,6 @@
+import { atom } from 'recoil';
+
+export const myListModalState = atom<'Edit' | 'Create' | 'Delete' | null>({
+  key: 'myListModal',
+  default: null,
+});
