+++ src/atoms/paramState.ts
@@ -1,0 +1,7 @@
+import { Params } from 'react-router-dom';
+import { atom } from 'recoil';
+
+export const paramState = atom<Params<string>>({
+  key: 'param',
+  default: {},
+});
