+++ src/atoms/showEditorState.ts
@@ -1,0 +1,6 @@
+import { atom } from 'recoil';
+
+export const showEditorState = atom({
+  key: 'showEditor',
+  default: false,
+});
